client: Request active community from Serlo API
