--- packages/public/client/src/feature-prototypes/donation-profiles/active-authors.ts
@@ -1,53 +1,0 @@
-/**
- * This file is part of Serlo.org.
- *
- * Copyright (c) 2013-2020 Serlo Education e.V.
- *
- * Licensed under the Apache License, Version 2.0 (the "License")
- * you may not use this file except in compliance with the License
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * @copyright Copyright (c) 2013-2020 Serlo Education e.V.
- * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
- * @link      https://github.com/serlo-org/serlo.org for the canonical source repository
- */
-export const activeAuthors = [
-  '73435',
-  '163773',
-  '146133',
-  '154641',
-  '80545',
-  '142379',
-  '87602',
-  '18981',
-  '27693',
-  '266',
-  '70932',
-  '64900',
-  '162709',
-  '146443',
-  '146667',
-  '122345',
-  '34105',
-  '164226',
-  '172113',
-  '132639',
-  '161065',
-  '167731',
-  '168463',
-  '150777',
-  '161064',
-  '167742',
-  '167651',
-  '141407',
-  '53589',
-  '98366',
-]
