--- packages/public/client/src/feature-prototypes/donation-profiles/active-donors.ts
@@ -1,57 +1,0 @@
-/**
- * This file is part of Serlo.org.
- *
- * Copyright (c) 2013-2020 Serlo Education e.V.
- *
- * Licensed under the Apache License, Version 2.0 (the "License")
- * you may not use this file except in compliance with the License
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * @copyright Copyright (c) 2013-2020 Serlo Education e.V.
- * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
- * @link      https://github.com/serlo-org/serlo.org for the canonical source repository
- */
-export const activeDonors = [
-  '27693',
-  '53589',
-  '10',
-  '396',
-  '2543',
-  '72550',
-  '22573',
-  '45848',
-  '51236',
-  '26217',
-  '70945',
-  '35408',
-  '73437',
-  '80545',
-  '108729',
-  '299',
-  '67608',
-  '64302',
-  '122345',
-  '49548',
-  '266',
-  '70932',
-  '73435',
-  '324',
-  '42099',
-  '40928',
-  '52009',
-  '1',
-  '49524',
-  '52733',
-  '270',
-  '81565',
-  '146133',
-  '65093',
-]
