--- packages/public/client/src/feature-prototypes/donation-profiles/active-reviewers.ts
@@ -1,35 +1,0 @@
-/**
- * This file is part of Serlo.org.
- *
- * Copyright (c) 2013-2020 Serlo Education e.V.
- *
- * Licensed under the Apache License, Version 2.0 (the "License")
- * you may not use this file except in compliance with the License
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * @copyright Copyright (c) 2013-2020 Serlo Education e.V.
- * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
- * @link      https://github.com/serlo-org/serlo.org for the canonical source repository
- */
-export const activeReviewers = [
-  '266',
-  '18981',
-  '27693',
-  '63374',
-  '70932',
-  '73435',
-  '73437',
-  '87602',
-  '105942',
-  '122345',
-  '146133',
-  '154641',
-]
